pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend logic for resume evaluator"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  # Web framework
  "fastapi==0.127.0",
  "mangum==0.20.0",

  # Config / env
  "python-dotenv==1.2.1",

  # Typing / validation
  "pydantic==2.11.10",
  "typing-extensions==4.15.0",

  # LangGraph / LangChain core
  "langgraph==1.0.5",
  "langgraph-checkpoint==3.0.1",
  "langchain-core==1.2.5",
  "langchain-openai==1.1.6",
  "langchain-community==0.4.1",

  # Tracing (optional, but imported)
  "langsmith==0.5.1",

  "google-search-results>=2.4.2",
]


[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
